+++ src/eve-server/system/cosmicMgrs/ManagerDB.cpp
@@ -149,7 +149,7 @@ GPoint ManagerDB::GetAnomalyPos(std::string& string)
         return NULL_ORIGIN;
     }
 
-    return (GPoint)(row.GetFloat(0), row.GetFloat(1), row.GetFloat(2));
+    return (GPoint)(row.GetDouble(0), row.GetDouble(1), row.GetDouble(2));
 }
 
 void ManagerDB::GetSystemAnomalies(uint32 systemID, DBQueryResult& res)
